+++ src/main/java/madstax/controller/listener/ModalEditorListener.java
@@ -1,4 +1,4 @@
-package madstax.controller;
+package madstax.controller.listener;
 
 import java.util.EventListener;
 import java.util.List;
